+++ Test/Models/Interface/Customer/INewCustomersCompanyRepo.cs
@@ -8,6 +8,6 @@ namespace Test.Models.Interface.Customer
 {
     public interface INewCustomersCompanyRepo: IDisposable
     {
-        List<usp_NewCustomersCompany_Result> usp_NewCustomersCompany(DateTime start, DateTime end, int companyRef, int timeType);
+        List<usp_NewCustomersCompany_Result> usp_NewCustomersCompany(DateTime start, DateTime end, long companyRef, int timeType);
     }
 }
