+++ Test/Models/Interface/Financial/IIncomeByProductCategoryCompanyRepo.cs
@@ -9,6 +9,6 @@ namespace Test.Models.Interface.Financial
 {
     public interface IIncomeByProductCategoryCompanyRepo : IDisposable
     {
-        List<usp_IncomeByProductCategoryCompany_Result> usp_IncomeByProductCategoryCompany(DateTime start, DateTime end, int companyRef);
+        List<usp_IncomeByProductCategoryCompany_Result> usp_IncomeByProductCategoryCompany(DateTime start, DateTime end, long companyRef);
     }
 }
