+++ Test/Models/Interface/Financial/ITotalSalesBranchRepo.cs
@@ -8,6 +8,6 @@ namespace Test.Models.Interface.Financial
 {
     public interface ITotalSalesBranchRepo : IDisposable
     {
-        List<usp_TotalSalesBranch_Result> usp_TotalSalesBranch(DateTime? start, DateTime? end, int? companyRef, int? branchRef, int? timeType); 
+        List<usp_TotalSalesBranch_Result> usp_TotalSalesBranch(DateTime? start, DateTime? end, long? companyRef, long? branchRef, int? timeType); 
     }
 }
